After uploading the sound the browser goes to the next probe.
Before: we sent and XMLHttpRequest (xhr) to probe/upload/ which returned
a redirect. The browser would follow the redirect in the background
with xhr.responseText containing the HTML of the requested page in the
end. Then we did document.write(xhr.responseText).

What we wanted: an actual redirect just like when you submit a form.

Solution: probe/upload/ now does not return a redirect but a response
with the redirect URL in the body. Then we just do
window.location = xhr.responseText